set up invite steps
